requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
boto3==1.34.162
pydantic>=2.0,<3.0